fix: add getAllTokens to landing page
